+++ src/room/index.js
@@ -3,6 +3,7 @@ let roomLogic = {
     defense:     require('./defense'),
     resources:     require('./resources'),
     construction:     require('./construction'),
+    population:     require('./population'),
 }
 
 module.exports = roomLogic;
